move code block and add styling
