+++ website/src/pages/blog/BlogComponents/CodeBlock/CodeBlock.css
@@ -1,0 +1,3 @@
+.code-div {
+    width: 80%;
+}
